Delete setters
There is no use case where a user of the library has to use one of these
setters. The fields are only set the RdcAppiumSuite runner by the method
"configure".
Removing the setters also helps the user of the library because they
don't have to think about these methods anymore.